// src/theme/theme.interface.ts
export default interface ThunderUITheme {
  name: string

  error?: string
  warning?: string

  primaryLighter: string
  primaryLightest: string
  primaryLight: string
  primary: string
  primaryStrong: string
  primaryStronger: string

  neutralLightest: string
  neutralLighter: string
  neutralLight: string
  neutral: string
  neutralStrong: string
  neutralStronger: string

  shadowLight: string
  shadow: string
  shadowStrong: string
}